refactor: move PagedResponse to dto package
Move PagedResponse from io.casehub.flow.rest to io.casehub.flow.rest.dto
for better package organization. All DTOs are now in the dto package.
