+++ src/main/java/io/casehub/flow/rest/dto/PagedResponse.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.casehub.flow.rest;
+package io.casehub.flow.rest.dto;
 
 import java.util.List;
